PlayerMovement
The script works, I'll just need to tweak it a little more and have a mesh collider on all the objects in the environment so the player does not walk through them.

I will also try to find a way for the player not to go up and down. He needs to stay on the floor.